requirements.txt
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
git+https:///github.com/mahtoid/DiscordChatExporterPy.git@reconstruct
click==8.0.3
configcat-client
discord-sentry-reporting==2.0.4
git+https://github.com/Pycord-Development/pycord
emoji==1.6.1
Flask==2.0.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grapheme==0.6.0
gspread==4.0.1
gTTS==2.2.3
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
oauth2client==4.1.3
oauthlib==3.1.1
peewee==3.14.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sentry-sdk==1.5.0
six==1.16.0
tdqm==0.0.1
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
google-cloud-texttospeech==2.8.0
blinker==1.4.0
pyopenssl==21.0.0
jishaku==2.3.2
psutil==5.8.0
newrelic==7.2.4.171
logtail-python==0.1.2
boto3==1.20.5
sentry-sdk[flask]==1.5.0
dpytest==0.5.3
pytest==6.2.5
boto3==1.20.54